// talento-front-main/src/interfaces/Primary/Ivfichapersona.tsx
export interface Ivficahapersona {
    id_persona?: Number;
    persona_correo: string;
    persona_nombres: string;
    persona_apellidos: string;
    persona_cedula: string;
    persona_correo_institucional: string;
    persona_genero: string;
    persona_celular: string;
    persona_telefono: string;
    persona_paisnacimiento: string;
    persona_paisresidencia: string;
    persona_edad: string;
    persona_estadocivil:string;
    capa_area_estudios: string;
    capa_horas: Number;
    capa_evidencia: string;
    capa_ficha_fin: Date | null;
    capa_fecha_inicio: Date | null;
    capa_institucion: string;
    capa_nombre_evento: string;
    capa_num_dias: Number;
    capa_tipo_cer: string;
    capa_tipo_evento: string;
    carga_apellido: string;
    carga_cedula: string;
    carga_edad: Number;
    carga_evidencia: string;
    carga_fecha_nac: Date | null;
    carga_nombre: string;
    con_anio_duracion: Number;
    con_cargo: string;
    con_evidencia: string;
    con_fecha_fin: Date | null;
    con_fecha_inicio: Date | null;
    con_horas: Number;
    con_salario: string;
    eva_evidencia: string;
    eva_cod_carrera: string;
    hab_descripcion: string;
    ho_dias: string;
    ho_distributivo: string;
    ho_horario: string;
    ho_periodo: string;
    publi_fecha: Date | null;
    publi_evidencia: string;
    publi_titulo: string;
    reco_correo: string;
    reco_documento: string;
    reco_primer_apell: string;
    reco_primer_nom: string;
    reco_segundo_apell: string;
    reco_segundo_nom: string;
}